# openapi.yaml
openapi: 3.1.1
info:
  title: API Store
  version: 1.0.1 
  description: API WITH OPENAPI
tags:
  - name: Users
    description: Operations related to users
  - name: Products
    description: Operations related to products
servers:
  - url: http://localhost:3000/v1
    description: Version 1 de API
  - url: http://localhost:3000/v2
    description: Version 2 de API
components:
  schemas:
    HelloResponse:
      type: object
      properties:
        message:
          type: string
      example:
        message: "Hello from API Store!"
    Product:
      type: object
      required:
      - name
      - price
      - category
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 40
        description:
          type: string
          maxLength: 500
        price:
          type: number
          minimum: 0
          multipleOf: 0.01
        category:
          type: string
          enum: [Electronics, Clothing, Books, Food, Home]
        tags:
          type: array
          items:
            type: string
            minLength: 1
        inStock:
          type: boolean
        specifications:
          type: object
          additionalProperties:
            type: string
        rating:
          type: array
          items:
            type: object
            required:
            - score
            - comment
            properties:
              score:
                type: integer
                minimum: 1
                maximum: 5
              comment:
                type: string
                maxLength: 200
      example:
        name: "Auriculares inalámbricos"
        description: "Auriculares con cancelación de ruido y hasta 30 horas de batería."
        price: 99.99
        category: Electronics
        tags:
          - audio
          - wireless
          - noise-cancelling
        inStock: true
        specifications:
          batteryLife: "30h"
          connectivity: "Bluetooth 5.2"
          color: "Negro"
        rating:
          - score: 5
            comment: "Excelente sonido y comodidad."
          - score: 4
            comment: "Buena relación calidad-precio."
    AuthRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
      example:
        email: "maria.perez@example.com"
        password: "SuperSecret123"
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          example: Bearer
      example:
        accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        tokenType: "Bearer"
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Autenticación mediante JSON Web Token (JWT)."
security: []
paths:
  /hello:
    get:
      summary: Hello Word
      responses:
        '200':
          description: A Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HelloResponse'
              example:
                message: "Hola desde API Store"
  /auth/login:
    post:
      tags:
        - Users
      summary: Autenticar usuario y obtener JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Token JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                tokenType: "Bearer"
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: "Credenciales inválidas"
  /users:
    post:
      tags:
        - Users
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - name
              - age
              - email
              properties:
                name:
                  type: string
                  minLength: 2
                age:
                  type: integer
                  minimum: 18
                email:
                  type: string
                  format: email
              example:
                name: "María Pérez"
                age: 28
                email: "maria.perez@example.com"
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  age:
                    type: integer
                  email:
                    type: string
              example:
                id: 123
                name: "María Pérez"
                age: 28
                email: "maria.perez@example.com"
  /users/{id}:
    get:
      tags:
          - Users
      summary: Get user by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  age:
                    type: integer
                  email:
                    type: string
              example:
                id: 123
                name: "María Pérez"
                age: 28
                email: "maria.perez@example.com"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: "Usuario no encontrado"        
    post:
      tags:
          - Users
      summary: Update user information
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - name
              - age
              - email
              properties:
                name:
                  type: string
                  minLength: 2
                age:
                  type: integer
                  minimum: 18
                email:
                  type: string
                  format: email
              example:
                name: "María Pérez"
                age: 29
                email: "maria.perez@ejemplo.com"
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  age:
                    type: integer
                  email:
                    type: string
              example:
                id: 123
                name: "María Pérez"
                age: 29
                email: "maria.perez@ejemplo.com"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: "Usuario no encontrado"
  /productos:
    get:
      tags:
          - Products
      summary: Obtener todos los productos
      responses:
        '200':
          description: Lista de productos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
              example:
                - name: "Auriculares inalámbricos"
                  description: "Auriculares con cancelación de ruido y hasta 30 horas de batería."
                  price: 99.99
                  category: Electronics
                  tags:
                    - audio
                    - wireless
                    - noise-cancelling
                  inStock: true
                  specifications:
                    batteryLife: "30h"
                    connectivity: "Bluetooth 5.2"
                    color: "Negro"
                  rating:
                    - score: 5
                      comment: "Excelente sonido y comodidad."
                    - score: 4
                      comment: "Buena relación calidad-precio."
    post:
      tags:
          - Products
      summary: Crear un nuevo producto
      security:
        - JWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
            example:
              name: "Auriculares inalámbricos"
              description: "Auriculares con cancelación de ruido y hasta 30 horas de batería."
              price: 99.99
              category: Electronics
              tags:
                - audio
                - wireless
                - noise-cancelling
              inStock: true
              specifications:
                batteryLife: "30h"
                connectivity: "Bluetooth 5.2"
                color: "Negro"
              rating:
                - score: 5
                  comment: "Excelente sonido y comodidad."
      responses:
        '201':
          description: Producto creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
              example:
                name: "Auriculares inalámbricos"
                description: "Auriculares con cancelación de ruido y hasta 30 horas de batería."
                price: 99.99
                category: Electronics
                tags:
                  - audio
                  - wireless
                  - noise-cancelling
                inStock: true
                specifications:
                  batteryLife: "30h"
                  connectivity: "Bluetooth 5.2"
                  color: "Negro"
                rating:
                  - score: 5
                    comment: "Excelente sonido y comodidad."
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: "Datos de producto inválidos, revisa los campos requeridos."
  /productos/{id}:
    get:
      tags:
          - Products
      summary: Obtener producto por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Producto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
              example:
                name: "Auriculares inalámbricos"
                description: "Auriculares con cancelación de ruido y hasta 30 horas de batería."
                price: 99.99
                category: Electronics
                tags:
                  - audio
                  - wireless
                  - noise-cancelling
                inStock: true
                specifications:
                  batteryLife: "30h"
                  connectivity: "Bluetooth 5.2"
                  color: "Negro"
                rating:
                  - score: 5
                    comment: "Excelente sonido y comodidad."
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: "Producto no encontrado"
    put:
      tags:
          - Products
      summary: Actualizar producto
      security:
        - JWT: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
            example:
              name: "Auriculares inalámbricos"
              description: "Auriculares con cancelación de ruido y hasta 30 horas de batería."
              price: 99.99
              category: Electronics
              tags:
                - audio
                - wireless
                - noise-cancelling
              inStock: true
              specifications:
                batteryLife: "30h"
                connectivity: "Bluetooth 5.2"
                color: "Negro"
              rating:
                - score: 5
                  comment: "Excelente sonido y comodidad."
      responses:
        '200':
          description: Producto actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
              example:
                name: "Auriculares inalámbricos"
                description: "Auriculares con cancelación de ruido y hasta 30 horas de batería."
                price: 99.99
                category: Electronics
                tags:
                  - audio
                  - wireless
                  - noise-cancelling
                inStock: true
                specifications:
                  batteryLife: "30h"
                  connectivity: "Bluetooth 5.2"
                  color: "Negro"
                rating:
                  - score: 5
                    comment: "Excelente sonido y comodidad."
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: "Producto no encontrado"
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: "Datos de producto inválidos, revisa los campos requeridos."
    delete:
      tags:
          - Products
      summary: Eliminar producto
      security:
        - JWT: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Producto eliminado
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: "Producto no encontrado"